Add files via upload
This is the entire code as written, including taking values from both .csvs; plotting the graphs with baselining and baselining/smoothing; sorting ranges into bins; and the rudimentary peak assignment code. 

For integrating with your ui, I'd use all the code prior to the graphs, but if you want to use just the rampy code and adjust, this is up to you.